+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: tsoares- <tsoares-@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/18 08:11:33 by tsoares-          #+#    #+#             */
-/*   Updated: 2025/01/21 21:31:57 by jgils            ###   ########.fr       */
+/*   Updated: 2025/01/23 09:01:44 by jgils            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -171,6 +171,7 @@ t_tab					*get_cmdtable(char **input, t_env *env);
 void					identify_tokens(t_tab *cmdtable);
 void					get_type(t_node *token);
 void					get_redirect_type(t_node *token);
+t_tab					*init_cmdtab(char **input);
 
 // TOKENIZER_UTILS.C ----------------------------------------------------------
 int						cmdlen(char **input);
@@ -178,6 +179,9 @@ void					add_node(t_node *newnode, t_list *list);
 void					add_list(t_list *newnode, t_tab *cmdtable);
 t_node					*new_node(char *token);
 t_list					*new_list(char **input);
+
+// PARSER.C ------------------------------------------------------------------
+t_tab					*get_cmdtable(char **input, t_env *env);
 
 // TOKEN_FORMAT.C ------------------------------------------------------------
 void					rm_quote(t_node *token);
